new routes for comment,like and tweet are added

+++ src/controllers/playlist.controller.js
@@ -2,6 +2,7 @@ import { asyncHandler } from "../utils/asyncHandler.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { Playlist } from "../models/playlist.model.js";
+import mongoose, { isValidObjectId } from "mongoose";
 
 const createPlaylist = asyncHandler(async (req, res) => {
     const { name, description } = req.body;
@@ -36,6 +37,11 @@ const getUserPlaylists = asyncHandler(async (req, res) => {
     if (!userId) {
         throw new ApiError(400, "userId is required")
     }
+
+    if (!mongoose.isValidObjectId(userId)) {
+        throw new ApiError(400, "user id is incorrect")
+    }
+
     const userPlaylist = await Playlist.find({ owner: userId })
 
     if (!userPlaylist || userPlaylist.length === 0) {
@@ -54,6 +60,10 @@ const getPlaylistById = asyncHandler(async (req, res) => {
         throw new ApiError(400, "Playlist Id is required")
     }
 
+    if (!mongoose.isValidObjectId(playlistId)) {
+        throw new ApiError(400, "playlist id is incorrect")
+    }
+
     const playlist = await Playlist.findById(playlistId);
 
     if (!playlist) {
@@ -69,6 +79,10 @@ const addVideoToPlaylist = asyncHandler(async (req, res) => {
     const { playlistId, videoId } = req.params
     if (!playlistId || !videoId) {
         throw new ApiError(400, "playlist and Video Id's are required")
+    }
+
+    if (!mongoose.isValidObjectId(playlistId) || !mongoose.isValidObjectId(userId)) {
+        throw new ApiError(400, "either playlist id or user id is incorrect")
     }
 
     const playlist = await Playlist.findByIdAndUpdate(playlistId, { $push: { videos: videoId } }, { new: true });
@@ -89,6 +103,10 @@ const removeVideoFromPlaylist = asyncHandler(async (req, res) => {
         throw new ApiError(400, "playlist and Video Id's are required")
     }
 
+    if (!mongoose.isValidObjectId(playlistId) || !mongoose.isValidObjectId(userId)) {
+        throw new ApiError(400, "either playlist id or user id is incorrect")
+    }
+
     const playlist = await Playlist.findByIdAndUpdate(playlistId, { $pull: { videos: videoId } }, { new: true });
 
     if (!playlist) {
@@ -106,6 +124,10 @@ const deletePlaylist = asyncHandler(async (req, res) => {
         throw new ApiError(400, "Playlist id is required")
     }
 
+    if (!mongoose.isValidObjectId(playlistId)) {
+        throw new ApiError(400, "playlist id is incorrect")
+    }
+
     const playlist = await Playlist.findByIdAndDelete(playlistId);
     if (!playlist) {
         throw new ApiError(404, "No playlist found")
@@ -121,6 +143,11 @@ const updatePlaylist = asyncHandler(async (req, res) => {
     if (!playlistId) {
         throw new ApiError(400, "Playlist id is required")
     }
+
+    if (!mongoose.isValidObjectId(playlistId)) {
+        throw new ApiError(400, "playlist id is incorrect")
+    }
+
     let updatesForPlaylist = {}
     if (name) {
         updatesForPlaylist.name = name;
